requirements.txt
alembic==1.13.1
blinker==1.8.2
Brotli==1.1.0
cachetools>=5.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cssselect2==0.7.0
email_validator>=2.0.0
Flask==3.0.2
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.54.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.6
MarkupSafe==2.1.5
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyodbc==5.2.0
pyphen==0.15.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
zopfli==0.2.3.post1